+++ jax/experimental/pallas/triton.py
@@ -14,10 +14,19 @@
 
 """Triton-specific Pallas APIs."""
 
+from jax._src.pallas.primitives import atomic_add as atomic_add
+from jax._src.pallas.primitives import atomic_and as atomic_and
+from jax._src.pallas.primitives import atomic_cas as atomic_cas
+from jax._src.pallas.primitives import atomic_max as atomic_max
+from jax._src.pallas.primitives import atomic_min as atomic_min
+from jax._src.pallas.primitives import atomic_or as atomic_or
+from jax._src.pallas.primitives import atomic_xchg as atomic_xchg
+from jax._src.pallas.primitives import atomic_xor as atomic_xor
 from jax._src.pallas.triton.core import CompilerParams as CompilerParams
 from jax._src.pallas.triton.primitives import approx_tanh as approx_tanh
 from jax._src.pallas.triton.primitives import debug_barrier as debug_barrier
 from jax._src.pallas.triton.primitives import elementwise_inline_asm as elementwise_inline_asm
+
 
 import typing as _typing  # pylint: disable=g-import-not-at-top
 if _typing.TYPE_CHECKING: