+++ docs/pallas/CHANGELOG.md
@@ -13,6 +13,13 @@ Remember to align the itemized text with the first line of an item within a list
 
 ## Unreleased
 
+* Deprecations
+
+  * `pl.atomic_*` APIs have been moved to {mod}`jax.experimental.pallas.triton`.
+    Accessing them via {mod}`jax.experimental.pallas` is deprecated.
+
+## Released with jax 0.7.0
+
 * New functionality
 
   * Added a new decorator {func}`jax.experimental.pallas.loop` which allows